prevent overwriting store conversation ids with 0

+++ app/msg/msg.model.php
@@ -561,23 +561,28 @@ class MsgModel extends Model
 
   public function setConversationMembers($cid,$fsids,$unread = False)
   {
-    $ur = 0;
-    if($unread)
-      $ur = 1;
-
-    if(count($fsids) < 1) {
-      $this->del('DELETE FROM `'.PREFIX.'foodsaver_has_conversation` WHERE conversation_id = '.(int)$cid);
-    } else {
-      $ids = implode(',', $fsids);
-      $this->del('DELETE FROM `'.PREFIX.'foodsaver_has_conversation` WHERE conversation_id = '.(int)$cid.' AND foodsaver_id NOT IN ('.$ids.')');
-      $values = array();
-      foreach($fsids as $user)
-      {
-        $values[] = '('.(int)$cid.', '.(int)$user.', '.$ur.')';
-      }
-      if(count($values) > 0)
-        $this->insert('INSERT IGNORE INTO `'.PREFIX.'foodsaver_has_conversation` (conversation_id, foodsaver_id, unread) VALUES '.implode(",",$values) );
-    }
+  	if((int)$cid > 0)
+  	{
+  		$ur = 0;
+  		if($unread)
+  			$ur = 1;
+  		
+  		if(count($fsids) < 1) {
+  			$this->del('DELETE FROM `'.PREFIX.'foodsaver_has_conversation` WHERE conversation_id = '.(int)$cid);
+  		} else {
+  			$ids = implode(',', $fsids);
+  			$this->del('DELETE FROM `'.PREFIX.'foodsaver_has_conversation` WHERE conversation_id = '.(int)$cid.' AND foodsaver_id NOT IN ('.$ids.')');
+  			$values = array();
+  			foreach($fsids as $user)
+  			{
+  				$values[] = '('.(int)$cid.', '.(int)$user.', '.$ur.')';
+  			}
+  			if(count($values) > 0)
+  				$this->insert('INSERT IGNORE INTO `'.PREFIX.'foodsaver_has_conversation` (conversation_id, foodsaver_id, unread) VALUES '.implode(",",$values) );
+  		}
+  		
+  		return $this->updateConversationMembers($cid);
+  	}
   }
 
 
@@ -585,9 +590,12 @@ class MsgModel extends Model
   {
     $ur = 0;
     if($unread)
-      $ur = 1;
+    {
+    	$ur = 1;
+    }
 
     $this->insert('INSERT IGNORE INTO `'.PREFIX.'foodsaver_has_conversation` (conversation_id, foodsaver_id, unread) VALUES ('.(int)$cid.', '.(int)$fsid.', '.$ur.')');
+  	return $this->updateConversationMembers($cid);
   }
 	
 	public function deleteUserFromConversation($cid,$fsid,$deleteAlways = False)
@@ -598,24 +606,31 @@ class MsgModel extends Model
 		if($deleteAlways || ((int)$this->qOne('SELECT COUNT(foodsaver_id) FROM `'.PREFIX.'foodsaver_has_conversation` WHERE conversation_id = '.(int)$cid) > 2))
 		{
 			$this->del('DELETE FROM `'.PREFIX.'foodsaver_has_conversation` WHERE conversation_id = '.(int)$cid.' AND foodsaver_id = '.(int)$fsid);
-			if($member = $this->qOne('SELECT member FROM '.PREFIX.'conversation WHERE id = '.(int)$cid))
-			{
-				$member = unserialize($member);
-				$out = array();
-				
-				foreach ($member as $k => $v)
-				{
-					if($v['id'] != fsId())
-					{
-						$out[$k] = $v;
-					}
-				}
-
-				return $this->update('UPDATE '.PREFIX.'conversation SET member = '.$this->strval(serialize($out)).' WHERE id = '.(int)$cid);
-			}
-			
-			return false;
+			return $this->updateConversationMembers($cid);
 		}
+		
+		return false;
+	}
+	
+	public function updateConversationMembers($cid)
+	{
+		if($member = $this->qOne('SELECT member FROM '.PREFIX.'conversation WHERE id = '.(int)$cid))
+		{
+			$member = unserialize($member);
+			$out = array();
+		
+			foreach ($member as $k => $v)
+			{
+				if((int)$v['id'] > 0)
+				{
+					$out[$k] = $v;
+				}
+			}
+		
+			return $this->update('UPDATE '.PREFIX.'conversation SET member = '.$this->strval(serialize($out)).' WHERE id = '.(int)$cid);
+		}
+			
+		return false;
 	}
 	
 	public function insertConversation($recipients,$locked = false)
